fix test failure
